pyproject.toml
[project]
name = "dotfiles"
version = "0.1.0"
description = "A unified dotfiles management system for development configurations"
authors = [
    {name = "Chad Walters"}
]
requires-python = ">=3.8"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "pyyaml>=6.0.1",
    "tqdm>=4.66.1",
    "rich>=13.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "ruff>=0.2.1"
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[project.scripts]
dotfiles = "dotfiles.cli:main"
